PSARC 2007/064 Unified POSIX and Windows Credentials for Solaris 4994017 data structure sharing between rpcbind and libnsl leads to accidents 6549510 Need the ability to store SIDs in the Solaris cred_t 6549515 PSARC 2007/064: uid_t and gid_t to become unsigned

+++ usr/src/cmd/bnu/ct.c
@@ -2,9 +2,8 @@
  * CDDL HEADER START
  *
  * The contents of this file are subject to the terms of the
- * Common Development and Distribution License, Version 1.0 only
- * (the "License").  You may not use this file except in compliance
- * with the License.
+ * Common Development and Distribution License (the "License").
+ * You may not use this file except in compliance with the License.
  *
  * You can obtain a copy of the license at usr/src/OPENSOLARIS.LICENSE
  * or http://www.opensolaris.org/os/licensing.
@@ -20,7 +19,7 @@
  * CDDL HEADER END
  */
 /*
- * Copyright 2005 Sun Microsystems, Inc.  All rights reserved.
+ * Copyright 2007 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 
@@ -532,8 +531,8 @@ int    code;
     if (*_Tty != '\0') {
 	CDEBUG(5, "chmod/chown %s\n", _Tty);
 	if (chown(_Tty , UUCPUID, TTYGID) < 0 ) {
-	    CDEBUG(5, "Can't chown to uid=%ld, ", (long) UUCPUID);
-	    CDEBUG(5, "gid=%ld\n", (long) TTYGID);
+	    CDEBUG(5, "Can't chown to uid=%u, ", UUCPUID);
+	    CDEBUG(5, "gid=%u\n", TTYGID);
 	}
 	if (chmod(_Tty , TTYMOD) < 0) {
 	    CDEBUG(5, "Can't chmod to %lo\n", (unsigned long) TTYMOD);
