+++ usr/src/cmd/getfacl/getfacl.c
@@ -2,9 +2,8 @@
  * CDDL HEADER START
  *
  * The contents of this file are subject to the terms of the
- * Common Development and Distribution License, Version 1.0 only
- * (the "License").  You may not use this file except in compliance
- * with the License.
+ * Common Development and Distribution License (the "License").
+ * You may not use this file except in compliance with the License.
  *
  * You can obtain a copy of the license at usr/src/OPENSOLARIS.LICENSE
  * or http://www.opensolaris.org/os/licensing.
@@ -20,7 +19,7 @@
  * CDDL HEADER END
  */
 /*
- * Copyright 2005 Sun Microsystems, Inc.  All rights reserved.
+ * Copyright 2007 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 
@@ -311,7 +310,7 @@ pruname(uid_t uid)
 	passwdp = getpwuid(uid);
 	if (passwdp == (struct passwd *)NULL) {
 		/* could not get passwd information: display uid instead */
-		(void) sprintf(uidp, "%ld", (long)uid);
+		(void) sprintf(uidp, "%u", uid);
 		return (uidp);
 	} else
 		return (passwdp->pw_name);
@@ -326,7 +325,7 @@ prgname(gid_t gid)
 	groupp = getgrgid(gid);
 	if (groupp == (struct group *)NULL) {
 		/* could not get group information: display gid instead */
-		(void) sprintf(gidp, "%ld", (long)gid);
+		(void) sprintf(gidp, "%u", gid);
 		return (gidp);
 	} else
 		return (groupp->gr_name);
